// react-frontend/src/api/models/feedback-period-statistics.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FeedbackPeriodStatistics
 */
export interface FeedbackPeriodStatistics {
    /**
     * 
     * @type {string}
     * @memberof FeedbackPeriodStatistics
     */
    'periodStart'?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackPeriodStatistics
     */
    'periodEnd'?: string;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'averageRating'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'totalFeedbackCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'positiveFeedbackCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'negativeFeedbackCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'neutralFeedbackCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackPeriodStatistics
     */
    'satisfactionRate'?: number;
}