// react-frontend/src/api/models/feedback-item-statistics.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FeedbackItemStatistics
 */
export interface FeedbackItemStatistics {
    /**
     * 
     * @type {string}
     * @memberof FeedbackItemStatistics
     */
    'menuItemName'?: string;
    /**
     * 
     * @type {number}
     * @memberof FeedbackItemStatistics
     */
    'feedbackCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackItemStatistics
     */
    'averageRating'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackItemStatistics
     */
    'minRating'?: number;
    /**
     * 
     * @type {number}
     * @memberof FeedbackItemStatistics
     */
    'maxRating'?: number;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof FeedbackItemStatistics
     */
    'ratingDistribution'?: { [key: string]: number; };
}